+++ src/Symfony/Component/ErrorHandler/Exception/FlattenException.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\ErrorRenderer\Exception;
+namespace Symfony\Component\ErrorHandler\Exception;
 
 use Symfony\Component\Debug\Exception\FlattenException as LegacyFlattenException;
 use Symfony\Component\HttpFoundation\Exception\RequestExceptionInterface;
@@ -25,7 +25,6 @@ use Symfony\Component\HttpKernel\Exception\HttpExceptionInterface;
  */
 class FlattenException extends LegacyFlattenException
 {
-    private $title;
     private $message;
     private $code;
     private $previous;
@@ -33,9 +32,11 @@ class FlattenException extends LegacyFlattenException
     private $traceAsString;
     private $class;
     private $statusCode;
+    private $statusText;
     private $headers;
     private $file;
     private $line;
+    private $asString;
 
     public static function create(\Exception $exception, $statusCode = null, array $headers = []): self
     {
@@ -60,12 +61,12 @@ class FlattenException extends LegacyFlattenException
         }
 
         if (class_exists(Response::class) && isset(Response::$statusTexts[$statusCode])) {
-            $title = Response::$statusTexts[$statusCode];
+            $statusText = Response::$statusTexts[$statusCode];
         } else {
-            $title = 'Whoops, looks like something went wrong.';
+            $statusText = 'Whoops, looks like something went wrong.';
         }
 
-        $e->setTitle($title);
+        $e->setStatusText($statusText);
         $e->setStatusCode($statusCode);
         $e->setHeaders($headers);
         $e->setTraceFromThrowable($exception);
@@ -171,14 +172,14 @@ class FlattenException extends LegacyFlattenException
         return $this;
     }
 
-    public function getTitle()
+    public function getStatusText()
     {
-        return $this->title;
+        return $this->statusText;
     }
 
-    public function setTitle(string $title): self
+    public function setStatusText(string $statusText): self
     {
-        $this->title = $title;
+        $this->statusText = $statusText;
 
         return $this;
     }
@@ -355,8 +356,19 @@ class FlattenException extends LegacyFlattenException
         return $this->traceAsString;
     }
 
+    public function setAsString(?string $asString)
+    {
+        $this->asString = $asString;
+
+        return $this;
+    }
+
     public function getAsString()
     {
+        if (null !== $this->asString) {
+            return $this->asString;
+        }
+
         $message = '';
         $next = false;