+++ src/Symfony/Component/HttpKernel/Tests/Controller/ErrorControllerTest.php
@@ -12,10 +12,8 @@
 namespace Symfony\Component\HttpKernel\Tests\Controller;
 
 use PHPUnit\Framework\TestCase;
-use Symfony\Component\ErrorRenderer\ErrorRenderer;
-use Symfony\Component\ErrorRenderer\ErrorRenderer\HtmlErrorRenderer;
-use Symfony\Component\ErrorRenderer\ErrorRenderer\JsonErrorRenderer;
-use Symfony\Component\ErrorRenderer\Exception\FlattenException;
+use Symfony\Component\ErrorHandler\ErrorRenderer\HtmlErrorRenderer;
+use Symfony\Component\ErrorHandler\Exception\FlattenException;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Controller\ErrorController;
@@ -28,12 +26,12 @@ class ErrorControllerTest extends TestCase
     /**
      * @dataProvider getInvokeControllerDataProvider
      */
-    public function testInvokeController(Request $request, FlattenException $exception, int $statusCode, string $content)
+    public function testInvokeController(Request $request, \Exception $exception, int $statusCode, string $content)
     {
         $kernel = $this->getMockBuilder(HttpKernelInterface::class)->getMock();
-        $errorRenderer = new ErrorRenderer([new HtmlErrorRenderer(), new JsonErrorRenderer()]);
+        $errorRenderer = new HtmlErrorRenderer();
         $controller = new ErrorController($kernel, null, $errorRenderer);
-        $response = $controller($request, $exception);
+        $response = $controller($exception);
 
         $this->assertSame($statusCode, $response->getStatusCode());
         self::assertStringContainsString($content, strtr($response->getContent(), ["\n" => '', '    ' => '']));
@@ -43,50 +41,23 @@ class ErrorControllerTest extends TestCase
     {
         yield 'default status code and HTML format' => [
             new Request(),
-            FlattenException::createFromThrowable(new \Exception()),
+            new \Exception(),
             500,
             'The server returned a "500 Internal Server Error".',
         ];
 
         yield 'custom status code' => [
             new Request(),
-            FlattenException::createFromThrowable(new NotFoundHttpException('Page not found.')),
+            new NotFoundHttpException('Page not found.'),
             404,
             'The server returned a "404 Not Found".',
-        ];
-
-        $request = new Request();
-        $request->attributes->set('_format', 'json');
-        yield 'custom format via _format attribute' => [
-            $request,
-            FlattenException::createFromThrowable(new \Exception('foo')),
-            500,
-            '{"title": "Internal Server Error","status": 500,"detail": "Whoops, looks like something went wrong."}',
-        ];
-
-        $request = new Request();
-        $request->headers->set('Accept', 'application/json');
-        yield 'custom format via Accept header' => [
-            $request,
-            FlattenException::createFromThrowable(new HttpException(405, 'Invalid request.')),
-            405,
-            '{"title": "Method Not Allowed","status": 405,"detail": "Whoops, looks like something went wrong."}',
-        ];
-
-        $request = new Request();
-        $request->headers->set('Content-Type', 'application/json');
-        yield 'custom format via Content-Type header' => [
-            $request,
-            FlattenException::createFromThrowable(new HttpException(405, 'Invalid request.')),
-            405,
-            '{"title": "Method Not Allowed","status": 405,"detail": "Whoops, looks like something went wrong."}',
         ];
 
         $request = new Request();
         $request->attributes->set('_format', 'unknown');
         yield 'default HTML format for unknown formats' => [
             $request,
-            FlattenException::createFromThrowable(new HttpException(405, 'Invalid request.')),
+            new HttpException(405, 'Invalid request.'),
             405,
             'The server returned a "405 Method Not Allowed".',
         ];
@@ -106,7 +77,7 @@ class ErrorControllerTest extends TestCase
                     $exception = $request->attributes->get('exception');
 
                     $this->assertSame($_controller, $request->attributes->get('_controller'));
-                    $this->assertInstanceOf(FlattenException::class, $exception);
+                    $this->assertInstanceOf(\Throwable::class, $exception);
                     $this->assertSame($code, $exception->getStatusCode());
                     $this->assertFalse($request->attributes->get('showException'));
 
@@ -116,7 +87,7 @@ class ErrorControllerTest extends TestCase
             )
             ->willReturn($response = new Response());
 
-        $controller = new ErrorController($kernel, $_controller, new ErrorRenderer([]));
+        $controller = new ErrorController($kernel, $_controller, new HtmlErrorRenderer());
 
         $this->assertSame($response, $controller->preview(new Request(), $code));
     }